[PATCH] knfsd nfs4 warning fixes

+++ fs/nfsd/nfs4proc.c
@@ -106,7 +106,8 @@ static inline int
 nfsd4_open(struct svc_rqst *rqstp, struct svc_fh *current_fh, struct nfsd4_open *open)
 {
 	int status;
-	dprintk("NFSD: nfsd4_open filename %.*s\n",open->op_fname.len, open->op_fname.data);
+	dprintk("NFSD: nfsd4_open filename %.*s\n",
+		(int)open->op_fname.len, open->op_fname.data);
 
 	/* This check required by spec. */
 	if (open->op_create && open->op_claim_type != NFS4_OPEN_CLAIM_NULL)
